ext api:

+++ src/main/java/com/example/practice/service/JournalEntryService.java
@@ -3,7 +3,10 @@ package com.example.practice.service;
 import com.example.practice.entity.JournalEntry;
 import com.example.practice.entity.User;
 import com.example.practice.repository.JournalEntryRepository;
+import jakarta.transaction.Transactional;
 import org.bson.types.ObjectId;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
@@ -19,12 +22,15 @@ public class JournalEntryService {
     @Autowired
     private UserService userService;
 
+    private static final Logger logger = LoggerFactory.getLogger(JournalEntryService.class);
+
+    @Transactional
     public void saveEntry(JournalEntry journalEntry, String username){
         User user = userService.findByUsername(username);
         journalEntry.setDate(LocalDateTime.now());
         JournalEntry saved = journalEntryRepository.save(journalEntry);
         user.getJournalEntries().add(saved);
-        userService.saveEntry(user);
+        userService.saveUser(user);
     }
     public List<JournalEntry> getAll(){
         return journalEntryRepository.findAll();
@@ -32,10 +38,23 @@ public class JournalEntryService {
     public Optional<JournalEntry> findById(ObjectId id){
         return journalEntryRepository.findById(id);
     }
-    public void deleteById(ObjectId id, String username){
-        User user = userService.findByUsername(username);
-        user.getJournalEntries().removeIf(x-> x.getId().equals(id));
-        userService.saveEntry(user);
-        journalEntryRepository.deleteById(id);
+    @Transactional
+    public Boolean deleteById(ObjectId id, String username){
+        try {
+            User user = userService.findByUsername(username);
+            Boolean removed = user.getJournalEntries().removeIf(x -> x.getId().equals(id));
+            if (removed) {
+                userService.saveUser(user);
+                journalEntryRepository.deleteById(id);
+            }
+            return removed;
+        }
+        catch (Exception e){
+            logger.info("Error in deleting for {}: ", username, e);
+            return false;
+        }
     }
+//    public List<JournalEntry> findByUsername(String username){
+//
+//    }
 }
